Split properties into a separate module.
