+++ core/src/avm1/globals/math.rs
@@ -1,4 +1,4 @@
-use crate::avm1::object::Attribute::*;
+use crate::avm1::property::Attribute::*;
 use crate::avm1::return_value::ReturnValue;
 use crate::avm1::{Avm1, Error, Object, UpdateContext, Value};
 use gc_arena::{GcCell, MutationContext};
